+++ application/controllers/Welcome.php
@@ -23,7 +23,7 @@ class Welcome extends Application
 		$baseurl = base_url();
 		$this->data = array(
 			'urlLink' => $baseurl
-		);
+    );
 		$this->data['pagebody'] = 'welcome_message';
 		$this->render();
 	}