+++ application/models/Accessories.php
@@ -12,4 +12,15 @@ class Accessories extends CSV_Model
 	function __construct() {
       parent::__construct('assets/csv/Accessories.csv','AccessoriesId');
     }
+
+    public function rules()
+    {
+        $config = array(
+            ['field' => 'Name', 'label' => 'Accessory Name', 'rules' => 'alpha_numeric_spaces|max_length[64]'],
+            ['field' => 'Strength', 'label' => 'Strength', 'rules' => 'integer|less_than[101]'],
+            ['field' => 'Dexterity', 'label' => 'Dexterity', 'rules' => 'integer|less_than[101]'],
+            ['field' => 'Intelligence', 'label' => 'Intelligence', 'rules' => 'integer|less_than[101]'],
+        );
+        return $config;
+    }
 }