+++ application/views/category_edit.php
@@ -1,0 +1,7 @@
+<h1>Category # {CategoryId}</h1>
+<form role="form" action="/Category/submit" method="post">
+    {fcategory}
+    <input type="submit" value="Update Category Name"/>
+</form>
+{error}
+<a href="/Category/cancel"><input type="button" value="Cancel the current edit"/></a>