+++ application/views/itemedit.php
@@ -1,0 +1,10 @@
+<h1>Accessory # {AccessoriesId}</h1>
+<form role="form" action="/Catalog/submit" method="post">
+    {faccessory}
+    {fstrength}
+    {fdexterity}
+    {fintelligence}
+    {zsubmit}
+</form>
+{error}
+<a href="/Catalog/cancel"><input type="button" value="Cancel the current edit"/></a>